+++ app/src/main/java/de/baumann/browser/unit/HelperUnit.java
@@ -126,7 +126,7 @@ public class HelperUnit {
         }
     }
 
-    public static void saveAs(final Activity activity, final String url) {
+    public static void saveAs(final Activity activity, final String url, final String name) {
 
         try {
             MaterialAlertDialogBuilder builder = new MaterialAlertDialogBuilder(activity);
@@ -143,10 +143,11 @@ public class HelperUnit {
             editTop.setHint(activity.getString(R.string.dialog_title_hint));
             editBottom.setHint(activity.getString(R.string.dialog_extension_hint));
 
-            String filename = URLUtil.guessFileName(url, null, null);
-            editTop.setText(HelperUnit.fileName(url));
-
+            String filename = name != null ? name : URLUtil.guessFileName(url, null, null);
             String extension = filename.substring(filename.lastIndexOf("."));
+            String prefix = filename.substring(0, filename.lastIndexOf("."));
+
+            editTop.setText(prefix);
             if (extension.length() <= 8) editBottom.setText(extension);
 
             builder.setView(dialogView);
@@ -168,7 +169,7 @@ public class HelperUnit {
                 String extension1 = editBottom.getText().toString().trim();
                 String filename1 = title + extension1;
 
-                if (title.isEmpty() || extension1.isEmpty() || !extension1.startsWith(".")) {
+                if (title.isEmpty() || !extension1.startsWith(".")) {
                     NinjaToast.show(activity, activity.getString(R.string.toast_input_empty));
                 } else {
                     if (BackupUnit.checkPermissionStorage(activity)) {
@@ -249,22 +250,37 @@ public class HelperUnit {
 
     public static void initTheme(Activity context) {
         sp = PreferenceManager.getDefaultSharedPreferences(context);
-        switch (Objects.requireNonNull(sp.getString("sp_theme", "1"))) {
-            case "2":
-                context.setTheme(R.style.AppTheme_day);
-                break;
-            case "3":
-                context.setTheme(R.style.AppTheme_night);
-                break;
-            case "4":
-                context.setTheme(R.style.AppTheme_wallpaper);
-                break;
-            case "5":
-                context.setTheme(R.style.AppTheme_OLED);
-                break;
-            default:
-                context.setTheme(R.style.AppTheme);
-                break;
+
+        if (sp.getBoolean("useDynamicColor", true)) {
+            switch (Objects.requireNonNull(sp.getString("sp_theme", "1"))) {
+                case "2":
+                    context.setTheme(R.style.AppTheme_wallpaper_day);
+                    break;
+                case "3":
+                    context.setTheme(R.style.AppTheme_wallpaper_night);
+                    break;
+                case "5":
+                    context.setTheme(R.style.AppTheme_OLED);
+                    break;
+                default:
+                    context.setTheme(R.style.AppTheme_wallpaper);
+                    break;
+            }
+        } else {
+            switch (Objects.requireNonNull(sp.getString("sp_theme", "1"))) {
+                case "2":
+                    context.setTheme(R.style.AppTheme_day);
+                    break;
+                case "3":
+                    context.setTheme(R.style.AppTheme_night);
+                    break;
+                case "5":
+                    context.setTheme(R.style.AppTheme_OLED);
+                    break;
+                default:
+                    context.setTheme(R.style.AppTheme);
+                    break;
+            }
         }
     }
 
@@ -356,7 +372,7 @@ public class HelperUnit {
             String extension1 = editBottom.getText().toString().trim();
             String filename1 = title + extension1;
 
-            if (title.isEmpty() || extension1.isEmpty() || !extension1.startsWith(".")) {
+            if (title.isEmpty() || !extension1.startsWith(".")) {
                 NinjaToast.show(activity, activity.getString(R.string.toast_input_empty));
             } else {
                 if (BackupUnit.checkPermissionStorage(activity)) {